+++ tests/testthat/test-module-deps-methods.R
@@ -211,13 +211,16 @@ test_that("3 levels of parent and child modules load and show correctly", {
     }
   } else {
     mySim <- simInit(modules = list("grandpar1"), paths = list(modulePath = tmpdir))
-    mg <- moduleGraph(mySim, FALSE)
-    expect_true(is(mg, "list"))
-    expect_true(is(mg$graph, "igraph"))
-    expect_true(is(mg$communities, "communities"))
-    expect_true(length(unique(mg$communities$member)) == 3)
-    expect_true(any(grepl("grandpar1", communities(mg$communities)[['1']])))
-    expect_true(identical(basename(communities(mg$communities)[['1']]),
-                          c("grandpar1", "par1", "par2", "child1", "child2")))
+    mg <- moduleGraph(mySim, FALSE) ## will be list if successful; NULL if not (no igraph GLPK support)
+    if (is(mg, "list")) {
+      expect_true(is(mg$graph, "igraph"))
+      expect_true(is(mg$communities, "communities"))
+      expect_true(length(unique(mg$communities$member)) == 3)
+      comm <- try(communities(mg$communities)[["1"]])
+      if (!is(comm, "try-error")) {
+        expect_true(any(grepl("grandpar1", comm)))
+        expect_true(identical(basename(comm), c("grandpar1", "par1", "par2", "child1", "child2")))
+      }
+    }
   }
 })