+++ tests/testthat/test-save.R
@@ -12,11 +12,10 @@ test_that("saving files (and memoryUse)", {
 
   origPlan <- future::plan()
   if (is(origPlan, "sequential"))
-    pl <- future::plan("multiprocess", workers = 2)
+    pl <- suppressWarnings(future::plan("multiprocess", workers = 2)) ## suppressed for checks in Rstudio
   on.exit({
     future::plan(origPlan)
   }, add = TRUE)
-
 
   times <- list(start = 0, end = 6, "month")
   parameters <- list(
@@ -90,7 +89,6 @@ test_that("saving files (and memoryUse)", {
   expect_true(file.exists(file.path(tmpdir, "landscape_month7.rds")))
   rm(mySim)
 
-
   # test when filename has a dot
   tmpdir <- paste0(tmpdir, ".sdfd.lkjlll")
 
@@ -115,7 +113,6 @@ test_that("saving files (and memoryUse)", {
   expect_true(file.exists(file.path(tmpdir, "caribou_month7.rds")))
   expect_true(file.exists(file.path(tmpdir, "landscape_month7.rds")))
   rm(mySim)
-
 })
 
 test_that("saving csv files does not work correctly", {
@@ -193,8 +190,8 @@ test_that("saveSimList does not work correctly", {
   expect_true(is.numeric(sim$landscape$DEM[]))
 
   # Now put it back to disk for subsequent test
-  sim$landscape <- writeRaster(sim$landscape, filename = tmpfile[1], overwrite = TRUE)
-
+  unlink(c(tmpfile[1], extension(tmpfile[1], "gri"))) ## needed because of hardlink shenanigans
+  sim$landscape <- writeRaster(sim$landscape, filename = tmpfile[1])
   mySim$landscape <- setMinMax(mySim$landscape)
   expect_true(all.equal(mySim, sim, check.environment = FALSE))
 
@@ -214,8 +211,8 @@ test_that("saveSimList does not work correctly", {
   file.remove(dir(dirname(tmpfile[1]), pattern = ".gr", full.names = TRUE))
   # rm(mySim)
 
-  assign("a", 1, envir = mySim@.xData$.mods$caribouMovement$mod)
-  assign("a", 2, envir = sim@.xData$.mods$caribouMovement$mod)
+  assign("a", 1, envir = mySim@.xData$.mods$caribouMovement$.objects)
+  assign("a", 2, envir = sim@.xData$.mods$caribouMovement$.objects)
 
   expect_true(bindingIsActive("mod", sim@.xData$.mods$caribouMovement))
   # test file-backed raster is gone
@@ -265,8 +262,6 @@ test_that("restart does not work correctly", {
     mySim <- simInit(times = times, params = parameters, modules = modules, paths = paths,
                      outputs = data.frame(objectName = "landscape", saveTime = times$end))
     mySim <- spades(mySim, debug = 1)
-
-
   } else if (testNum == 2) {
     options("spades.restartRInterval" = 10)
     mySim <- mySim <- simInit(