+++ NEWS.md
@@ -10,6 +10,7 @@ version 1.0.5.9000
 * no changes
 
 ## bug fixes
+* use `try()` with `communities()` to skip tests on systems without `igraph` GLPK support.
 * minor bug fixes
 
 version 1.0.5