+++ NEWS.md
@@ -11,6 +11,7 @@ version 1.0.5.9000
 
 ## bug fixes
 * use `try()` with `communities()` to skip tests on systems without `igraph` GLPK support.
+* prevent package (re)installation during examples, tests, vignettes.
 * minor bug fixes
 
 version 1.0.5