+++ NEWS.md
@@ -1,5 +1,19 @@
 Known issues: https://github.com/PredictiveEcology/SpaDES.core/issues
 
+version 1.0.6
+=============
+
+## new features
+* none
+
+## dependencies
+* no changes
+
+## bug fixes
+* use `try()` with `communities()` to skip tests on systems without `igraph` GLPK support.
+* prevent package (re)installation during examples, tests, vignettes.
+* minor bug fixes
+
 version 1.0.5
 =============